+++ apps/backend/src/modules/project-groups/project-group.controller.ts
@@ -1,4 +1,5 @@
-import { APIPath } from "~/libs/enums/enums.js";
+import { APIPath, PermissionKey } from "~/libs/enums/enums.js";
+import { checkUserPermissions } from "~/libs/hooks/hooks.js";
 import {
 	type APIHandlerOptions,
 	type APIHandlerResponse,
@@ -12,8 +13,12 @@ import { ProjectGroupsApiPath } from "./libs/enums/enums.js";
 import {
 	type ProjectGroupCreateRequestDto,
 	type ProjectGroupGetAllRequestDto,
+	type ProjectGroupPatchRequestDto,
 } from "./libs/types/types.js";
-import { projectGroupCreateValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
+import {
+	projectGroupCreateValidationSchema,
+	projectGroupPatchValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";
 import { type ProjectGroupService } from "./project-group.service.js";
 
 /**
@@ -87,6 +92,22 @@ class ProjectGroupController extends BaseController {
 				),
 			method: "GET",
 			path: ProjectGroupsApiPath.$ID,
+			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_ALL_PROJECTS])],
+		});
+
+		this.addRoute({
+			handler: (options) =>
+				this.patch(
+					options as APIHandlerOptions<{
+						body: ProjectGroupPatchRequestDto;
+						params: { id: string };
+					}>,
+				),
+			method: "PATCH",
+			path: ProjectGroupsApiPath.$ID,
+			validation: {
+				body: projectGroupPatchValidationSchema,
+			},
 		});
 	}
 
@@ -195,6 +216,49 @@ class ProjectGroupController extends BaseController {
 			payload: await this.projectGroupService.findAllByProjectId(
 				params.id,
 				query,
+			),
+			status: HTTPCode.OK,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /groups/{id}:
+	 *    patch:
+	 *      tags:
+	 *        - ProjectGroup
+	 *      description: Patch project group info
+	 *      parameters:
+	 *        - in: path
+	 *          name: id
+	 *          description: ID of the project group to patch
+	 *          schema:
+	 *            type: string
+	 *      responses:
+	 *        200:
+	 *          description: Successful operation
+	 *          content:
+	 *            application/json:
+	 *              schema:
+	 *                type: object
+	 *                properties:
+	 *                  message:
+	 *                    type: object
+	 *                    $ref: "#/components/schemas/ProjectGroup"
+	 */
+
+	private async patch(
+		options: APIHandlerOptions<{
+			body: ProjectGroupPatchRequestDto;
+			params: { id: string };
+		}>,
+	): Promise<APIHandlerResponse> {
+		const projectGroupId = Number(options.params.id);
+
+		return {
+			payload: await this.projectGroupService.patch(
+				projectGroupId,
+				options.body,
 			),
 			status: HTTPCode.OK,
 		};