+++ apps/frontend/src/libs/constants/navigation-items.constant.ts
@@ -6,7 +6,10 @@ const SIDEBAR_ITEMS: NavigationItem[] = [
 		href: AppRoute.PROJECTS,
 		icon: "project",
 		label: "Projects",
-		pagePermissions: [PermissionKey.VIEW_ALL_PROJECTS],
+		pagePermissions: [
+			PermissionKey.VIEW_ALL_PROJECTS,
+			PermissionKey.MANAGE_ALL_PROJECTS,
+		],
 	},
 	{
 		href: AppRoute.ACCESS_MANAGEMENT,