+++ apps/frontend/src/pages/project-access-management/libs/components/project-group-form/project-group-form.tsx
@@ -11,33 +11,37 @@ import {
 	useSelectedItems,
 } from "~/libs/hooks/hooks.js";
 import { type ValidationSchema } from "~/libs/types/types.js";
-import { type ProjectGroupCreateRequestDto } from "~/modules/project-groups/project-groups.js";
+import {
+	type ProjectGroupCreateRequestDto,
+	type ProjectGroupPatchRequestDto,
+} from "~/modules/project-groups/project-groups.js";
 import { actions as projectPermissionActions } from "~/modules/project-permissions/project-permissions.js";
-import { type UserGetAllItemResponseDto } from "~/modules/users/users.js";
 
 import { ProjectGroupUsersTable } from "../project-group-users-table/project-group-users-table.js";
 import { getPermissionOptions } from "./libs/helpers/helpers.js";
 import styles from "./styles.module.css";
 
-type Properties<T extends ProjectGroupCreateRequestDto> = {
+type ProjectGroupRequest =
+	| ProjectGroupCreateRequestDto
+	| ProjectGroupPatchRequestDto;
+
+type Properties<T extends ProjectGroupRequest> = {
 	defaultValues: T;
 	onSubmit: (payload: T) => void;
 	submitLabel: string;
-	users: UserGetAllItemResponseDto[];
 	validationSchema: ValidationSchema<T>;
 };
 
-const ProjectGroupForm = <T extends ProjectGroupCreateRequestDto>({
+const ProjectGroupForm = <T extends ProjectGroupRequest>({
 	defaultValues,
 	onSubmit,
 	submitLabel,
-	users,
 	validationSchema,
 }: Properties<T>): JSX.Element => {
 	const dispatch = useAppDispatch();
 
 	const { control, errors, handleErrorsClear, handleSubmit, handleValueSet } =
-		useAppForm<ProjectGroupCreateRequestDto>({
+		useAppForm<ProjectGroupRequest>({
 			defaultValues,
 			validationSchema,
 		});
@@ -60,7 +64,7 @@ const ProjectGroupForm = <T extends ProjectGroupCreateRequestDto>({
 
 	const handleFormSubmit = useCallback(
 		(event_: React.BaseSyntheticEvent): void => {
-			void handleSubmit((formData: ProjectGroupCreateRequestDto) => {
+			void handleSubmit((formData: ProjectGroupRequest) => {
 				onSubmit(formData as T);
 			})(event_);
 		},
@@ -94,7 +98,6 @@ const ProjectGroupForm = <T extends ProjectGroupCreateRequestDto>({
 				onToggle={handleUserIdsToggle}
 				selectedUserIds={selectedUserIds}
 				setValue={handleValueSet}
-				users={users}
 			/>
 			<Select
 				control={control}