+++ apps/frontend/src/pages/project-access-management/libs/components/project-group-users-table/libs/components/project-group-users-search/project-group-users-search.tsx
@@ -1,0 +1,22 @@
+import { Search } from "~/libs/components/components.js";
+
+import styles from "./styles.module.css";
+
+type Properties = {
+	onChange: (search: string) => void;
+};
+
+const ProjectGroupUsersSearch = ({ onChange }: Properties): JSX.Element => {
+	return (
+		<div className={styles["search-container"]}>
+			<Search
+				isLabelHidden
+				label="Users search"
+				onChange={onChange}
+				placeholder="Enter name"
+			/>
+		</div>
+	);
+};
+
+export { ProjectGroupUsersSearch };