+++ scripts/analytics/src/libs/modules/analytics-cli/analytics-cli.ts
@@ -6,4 +6,3 @@ const analyticsCli = new BaseAnalyticsCli({
 });
 
 export { analyticsCli };
-export { getPackageName } from "./init-analytics-cli.js";